requirements.txt
streamlit
pandas
nltk
scikit-learn
google.generativeai